--- appium/api-mocks/apis/google/constants.ts
@@ -1,57 +1,0 @@
-const testCertificateMultipleSmimeCEA2D53BB9D24871 = `-----BEGIN CERTIFICATE-----
-MIIEUzCCAjugAwIBAgIBAjANBgkqhkiG9w0BAQUFADB0MRMwEQYDVQQIDApTb21l
-LVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQxFzAVBgNV
-BAMMDlNvbWUgQXV0aG9yaXR5MSEwHwYJKoZIhvcNAQkBFhJhdXRob3JpdHlAdGVz
-dC5jb20wIBcNMjEwOTI2MTAzOTUwWhgPMjEyMTA5MjYxMDM5NTBaMDAxLjAsBgNV
-BAMTJWZsb3djcnlwdC50ZXN0LmtleS5tdWx0aXBsZUBnbWFpbC5jb20wggEiMA0G
-CSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDaYbbw4ikp4aOLxmsUvVp0BD6id7dx
-K/ZbdC7rnyy1X/nkRAHEf3xrjPZk4KPPy8FGSgp/jIL1I+vEglcEPbBm61ZKdYFy
-Vty9kmWa0b6Ykf8urKR0hAfSCZheE5ua0OQ6cfmdOIZKJVRLLjAQql3MBGl4aWqn
-686sgoiazIxbgJ8HPi6I9GPSk3RSDHqksXqQUYoNL5Ugi2WGYhByf+BB5sS6vA0T
-gD+FkRnFGC56Y90Wkne3QofZP5/VPBeYV7bnxkJAl4I5yDNOMKrnBIL6pKBUqolj
-X6ARUFLCWF+yG1rMUBgGgT1GDx2S1+fgpSxRBpYmexfsesShA0nJy339AgMBAAGj
-MjAwMAwGA1UdEwQFMAMBAf8wCwYDVR0PBAQDAgL0MBMGA1UdJQQMMAoGCCsGAQUF
-BwMEMA0GCSqGSIb3DQEBBQUAA4ICAQDbBf3xb7esrnxEjwL/2dvnHAqgUTVaf6O4
-hW9mUp4c727fbDqat9OlilI/fLcwBN/NgEf+9jSWq9uHvjklCIXNTaK8zTkuQdtX
-FXZjT7N9EXWO93wUFtBY/fUIhJAEaL7Z6rmWUkK6UW/mtV7I6gxiZ2GAAE2wFKyS
-atb/Q+sgyeSOZjxux9NRMh8ZXmHTBQ/RUkzyLFM4Kgn5qfbM2vpkVuwjc4aH6r6F
-qeRKf1HULybJn/aHQ6zxVqNYodekUHe55DIAA71X2PVEaltN7szE4K+obdcE4y2I
-OW0w2H/SvBgngkn1aIvHIOltSBczvnSqKyQnLae4FW8PJJyk/X1HkAZPlMIRXXl9
-eUsv81oSiSK7YmRGuObgldBeDYmmECXnknsOPekK+mBaBbNGZNFMS6iJZV4anl4n
-UGgG49YEW97yFhJ4r7CxLhP2Eo7Fbfax4wUDPfaNLv9OK3zn27sjA/QSI3q7eBvi
-vn9dCnAk/hIBj/FsiyZstcczmvnHE9w9NDdCc0bbYcfwHAAs7XNejjDcenA+9ngq
-4hQ6ofK/4WIVekDJLRvoxNY94qjErT6iXL5aduYu8BVri2Q3Jqrp9YILLVH83rkl
-W8ZiWSb2P6bWZG3sxVNxS5iM6v39NbubnqK5UxE6+/GCeo+lHRNzZhhR3VRpKVw2
-xSncWuKSTg==
------END CERTIFICATE-----`;
-
-export const testConstants = {
-  testKeyMultipleSmimeCEA2D53BB9D24871: `-----BEGIN RSA PRIVATE KEY-----
-MIIEogIBAAKCAQEA2mG28OIpKeGji8ZrFL1adAQ+one3cSv2W3Qu658stV/55EQB
-xH98a4z2ZOCjz8vBRkoKf4yC9SPrxIJXBD2wZutWSnWBclbcvZJlmtG+mJH/Lqyk
-dIQH0gmYXhObmtDkOnH5nTiGSiVUSy4wEKpdzARpeGlqp+vOrIKImsyMW4CfBz4u
-iPRj0pN0Ugx6pLF6kFGKDS+VIItlhmIQcn/gQebEurwNE4A/hZEZxRguemPdFpJ3
-t0KH2T+f1TwXmFe258ZCQJeCOcgzTjCq5wSC+qSgVKqJY1+gEVBSwlhfshtazFAY
-BoE9Rg8dktfn4KUsUQaWJnsX7HrEoQNJyct9/QIDAQABAoIBAEWBklOK+CUPXxlu
-1sgsGG0SqMjXJ7jKkEe9a+2spPM5j/S6PIHJPdWRcR7cbOcHaAWyLTM5irjxbOoH
-viBobgvj3XwneELnKKzhupsTJQG0fi7h3hoDp+WPDkLrJLyavfloK1WCFyPkS7b+
-BWhYgTAdO2CK1NqLcCVyCn6Smh3I9LsA0z9VsMdPW4pKtQxNEuLMy3zb5Avy9RWH
-+/K3ysm/HHZi7yT84VKKzJEdCBqyEqIDEPibulbEZkQuBrvfrOWzz1lqI63ZDFkm
-by8XtyDh3Tox228Es4gdxmk4QCwEOU/51CuSCMvWAaUWd0ob+VS4mISCijLkbBCf
-nT41nQECgYEA9K+ln0pMBTgHD4hUhA3tIr1X8YtSl08FbZYg/dCPhq4/S0CnJJ0G
-h9G4V7yyNyYu0yZj+FK9ZC46j8oTKTFt6vB+gQLHG/HowE0df4MbIGCA/vpbsZdO
-/+2WUtfJ/ecKx9/89wHQkQIX95LWLu4P9+xncq3kp+YbVpBkXq2yPL0CgYEA5Hqz
-qtHokU2PGHecKTbm6dkwYNqg2wO4JuN3CwPIDUUJDbonTHBKJOOfoFhGbza5vc75
-k8HiMWUz29cFZv8d0wSpTI94yESAS5GqEp38+E/HI/jozGv0UUnd7Z4QP7IlXDIn
-2l7n0p6tPDmdQITfUK+PjI/6ow//1kX8uL54ekECgYA5um8S486HtK5Fxd5awYZf
-GdjzzfEQbb5UGoVyHJCgL+AS+w/0HW/6/0nEuWo6hLGrc3VGkw7H4fhDEGPw5g6O
-zWqrOWKMf0hwhkEdYms1+k602ZBkl9Q1oXJD3VF0q9vpSDlAdZclfx70dBpAPD5c
-OdQ2RDycODfe9nYXuFdC6QKBgGlebR2KUMcyDZwrDUm0mCmgkD8alLRssrC8llnu
-FISKIfLCtvz5jDgXAbbg/xqrEl4G5fLOp5JAKfpo34TwgsCelbVJNVqmvMQdGWIc
-ml2p7R0Z3cLoxnP2rPK4wqWdcr1iJB8BRchMhXgQ2gS3QoXjcUBR5jTtW/9lp3dL
-4CQBAoGAJPZ8pb1wmMohs4ZvcnbH0Y+uoMdoxQ+FLD4PzaPlRvZUQ8/PIBYyJ/f+
-lpiLc4Tbuv6u4a6thk6X8i/hFQQs9Uo6cil5Onj7llOh9YrmAqPAeiZTZLwHhd1L
-6VL6r+WR5hdFzNldGFQdZXJwUXArSmUHebw8QESneljBahWgKC0=
------END RSA PRIVATE KEY-----
-${testCertificateMultipleSmimeCEA2D53BB9D24871}`,
-};
